// generated/accounts/market.ts
export const MARKET_DISCRIMINATOR = new Uint8Array([
  219, 190, 213, 55, 0, 227, 198, 154,
]);

export function getMarketDiscriminatorBytes() {
  return fixEncoderSize(getBytesEncoder(), 8).encode(MARKET_DISCRIMINATOR);
}

export type Market = {
  discriminator: ReadonlyUint8Array;
  authority: Address;
  marketId: number;
  settlementDeadline: bigint;
  collateralMint: Address;
  collateralVault: Address;
  isSettled: boolean;
  winningOutcome: Option<WinningOutcome>;
  totalCollateralLocked: bigint;
  bump: number;
  metaDataUrl: string;
  outcomeYesMint: Address;
  outcomeNoMint: Address;
  yesEscrow: Address;
  noEscrow: Address;
};